+++ python/ray/workflow/exceptions.py
@@ -35,3 +35,20 @@ class WorkflowNotResumableError(WorkflowError):
     def __init__(self, workflow_id: str):
         self.message = f"Workflow[id={workflow_id}] is not resumable."
         super().__init__(self.message)
+
+
+@PublicAPI(stability="beta")
+class WorkflowNotFoundError(WorkflowError):
+    def __init__(self, workflow_id: str):
+        self.message = f"Workflow[id={workflow_id}] was referenced but doesn't exist."
+        super().__init__(self.message)
+
+
+@PublicAPI(stability="beta")
+class WorkflowRunningError(WorkflowError):
+    def __init__(self, operation: str, workflow_id: str):
+        self.message = (
+            f"{operation} couldn't be completed becasue "
+            f"Workflow[id={workflow_id}] is still running."
+        )
+        super().__init__(self.message)
